Adept our code for BIG ADRESSES

+++ src/debugger.cpp
@@ -4,8 +4,6 @@
 #include <iomanip>
 #include <fstream>
 #include <cstring>
-#include <data.hh>
-
 
 #include "linenoise/linenoise.h"
 #include "utility.h"
@@ -20,7 +18,7 @@ void debugger::initialise_load_address() {
         std::string addr;
         std::getline(map, addr, '-');
 
-        m_load_address = std::stoi(addr, 0, 16);
+        m_load_address = std::stol(addr, 0, 16);
     }
 }
